requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9

pillow==10.2.0
numpy<2
scikit-learn==1.4.1.post1

torch==2.1.2+cpu
torchvision==0.16.2+cpu

python-dotenv==1.0.1
